Remove include form sdk tests
